requirements.txt
asgiref==3.2.10
attrs==19.3.0
Django==3.0.7
django-classy-tags==1.0.0
django-cookie-law==2.0.3
django-crispy-forms==1.9.1
importlib-metadata==1.7.0
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
pytest-runner==5.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
wcwidth==0.2.5
zipp==3.1.0
